python_graphql2/requirements.txt
aniso8601==9.0.1
anyio==3.6.2
backports.cached-property==1.0.2
black==22.10.0
cffi==1.15.1
click==8.1.3
cryptography==41.0.6
fastapi==0.87.0
graphene==3.1.1
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
httptools==0.5.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
pycparser==2.21
pydantic==1.10.2
Pygments==2.15.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-gql==0.2.4
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.27.0
starlette-graphene3==0.6.0
strawberry-graphql==0.142.2
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
